mining: Move BgBlkTmplGenerator to separate file.
This moves BgBlkTmplGenerator and the related logic to a separate file,
mining/bgblktmplgenerator.go.

This is part of an effort to split mining into several files to make the
package easier to navigate, and in preparation of adding additional test
coverage.
